Moving all Spock tests into a separate project in extras
Replacing previous Spock tests with standard jUnit tests
