+++ pom.xml
@@ -130,6 +130,11 @@
                 <scope>test</scope>
             </dependency>
             <dependency>
+                <groupId>org.jboss.arquillian.container</groupId>
+                <artifactId>arquillian-container-test-api</artifactId>
+                <version>${org.jboss.arquillian.version}</version>
+            </dependency>
+            <dependency>
                 <groupId>org.javaee7</groupId>
                 <artifactId>util</artifactId>
                 <version>1.0-SNAPSHOT</version>
@@ -342,6 +347,7 @@
                 <artifactId>maven-war-plugin</artifactId>
                 <version>2.1.1</version>
                 <configuration>
+                    <attachClasses>true</attachClasses>
                     <failOnMissingWebXml>false</failOnMissingWebXml>
                 </configuration>
             </plugin>
